Ignore ndjson but run for each file in dir
